requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
supabase>=2.0.0
pandas>=2.0.0
requests>=2.31.0
schedule>=1.2.0
pytz>=2023.3
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
yfinance>=0.2.0
httpx>=0.25.0
google-cloud-aiplatform>=1.25.0
selenium>=4.15.0
pymongo>=4.6.0
motor>=3.3.0

